Using PyQt6 based gui and .venv
